graph more robust
fixed issue with graph points line where the line would go back on
itself effectively generating negative time. Also changed the way data
is saved so that there is less data bloat in the XML file. Also made one
or two touch ups such as changing the text size on the coordinate
system. Also refactored Graph code so that the update function works
with a Raycast and not RaycastAll function.